move get specific structs from lib to get
